+++ frontend/src/app/globals.css
@@ -86,13 +86,8 @@ body {
 }
 
 body {
-  color: rgb(var(--foreground-rgb));
-  background: linear-gradient(
-      to bottom,
-      transparent,
-      rgb(var(--background-end-rgb))
-    )
-    rgb(var(--background-start-rgb));
+  color: #333;
+  background: #fff
 }
 
 a {
@@ -102,6 +97,5 @@ a {
 
 @media (prefers-color-scheme: dark) {
   html {
-    color-scheme: dark;
   }
 }
